requirements.txt
# Polymarket HFT Bot - Dependencies
py-clob-client>=0.34.0
web3>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
binance-connector>=4.0.0
rich>=13.0.0
websocket-client>=1.6.0
